--- pyvcell/data_model/__init__.py
@@ -1,22 +1,0 @@
-from pyvcell.data_model.plotter import Plotter
-from pyvcell.data_model.result import Result
-from pyvcell.data_model.sbml_spatial_model import SbmlSpatialModel
-from pyvcell.data_model.simulation import SbmlSpatialSimulation, VcmlSpatialSimulation
-from pyvcell.data_model.var_types import NDArray1D, NDArray2D, NDArray3D
-from pyvcell.data_model.vcml_spatial_model import VcmlSpatialModel
-from pyvcell.data_model.vtk_data import VtkData
-from pyvcell.data_model.zarr_types import Channel
-
-__all__ = [
-    "Plotter",
-    "Result",
-    "SbmlSpatialModel",
-    "SbmlSpatialSimulation",
-    "VcmlSpatialSimulation",
-    "VcmlSpatialModel",
-    "VtkData",
-    "Channel",
-    "NDArray1D",
-    "NDArray2D",
-    "NDArray3D",
-]
