+++ tests/data_model/test_simulations.py
@@ -3,7 +3,10 @@ from pathlib import Path
 import numpy as np
 
 import pyvcell.vcml as vc
-from pyvcell.data_model import Result, SbmlSpatialModel, SbmlSpatialSimulation, VcmlSpatialModel, VcmlSpatialSimulation
+from pyvcell.data_model.result import Result
+from pyvcell.data_model.sbml_spatial_model import SbmlSpatialModel
+from pyvcell.data_model.simulation import SbmlSpatialSimulation, VcmlSpatialSimulation
+from pyvcell.data_model.vcml_spatial_model import VcmlSpatialModel
 
 
 def test_sbml_model_parse_1d(sbml_spatial_model_1d_path: Path) -> None:
