+++ tests/vcml/test_creation.py
@@ -1,7 +1,9 @@
 from pathlib import Path
 
 import pyvcell.vcml as vc
-from pyvcell.data_model import Result, VcmlSpatialModel, VcmlSpatialSimulation
+from pyvcell.data_model.result import Result
+from pyvcell.data_model.simulation import VcmlSpatialSimulation
+from pyvcell.data_model.vcml_spatial_model import VcmlSpatialModel
 
 
 def test_create() -> None:
